+++ Web/src/views/crm/clue/index.vue
@@ -4,12 +4,7 @@
 
   <ContentWrap>
     <!-- 搜索工作栏 -->
-    <el-form
-      class="-mb-15px"
-      :model="queryParams"
-      ref="queryFormRef"
-      label-width="auto"
-    >
+    <el-form class="-mb-15px" :model="queryParams" ref="queryFormRef" label-width="auto">
       <el-row :gutter="20">
         <el-col :span="8">
           <el-form-item :label="t('clue.name')" prop="name">
@@ -58,8 +53,12 @@
       <el-row>
         <el-col :span="24">
           <el-form-item>
-            <el-button @click="handleQuery"><Icon icon="ep:search" class="mr-5px" /> {{ t('common.search') }}</el-button>
-            <el-button @click="resetQuery"><Icon icon="ep:refresh" class="mr-5px" /> {{ t('common.reset') }}</el-button>
+            <el-button @click="handleQuery"
+              ><Icon icon="ep:search" class="mr-5px" /> {{ t('common.search') }}</el-button
+            >
+            <el-button @click="resetQuery"
+              ><Icon icon="ep:refresh" class="mr-5px" /> {{ t('common.reset') }}</el-button
+            >
             <el-button type="primary" @click="openForm('create')" v-hasPermi="['crm:clue:create']">
               <Icon icon="ep:plus" class="mr-5px" /> {{ t('action.add') }}
             </el-button>
@@ -84,9 +83,22 @@
       <el-tab-pane :label="t('customer.myResponsible')" name="1" />
       <el-tab-pane :label="t('customer.myInvolved')" name="2" />
       <el-tab-pane :label="t('customer.subordinateResponsible')" name="3" />
+      <el-tab-pane :label="t('customer.organizationScope')" name="4" />
     </el-tabs>
-    <el-table v-loading="loading" :data="list" :stripe="true" :show-overflow-tooltip="true" :table-layout="'auto'">
-      <el-table-column :label="t('clue.name')" align="center" prop="name" fixed="left" min-width="160">
+    <el-table
+      v-loading="loading"
+      :data="list"
+      :stripe="true"
+      :show-overflow-tooltip="true"
+      :table-layout="'auto'"
+    >
+      <el-table-column
+        :label="t('clue.name')"
+        align="center"
+        prop="name"
+        fixed="left"
+        min-width="160"
+      >
         <template #default="scope">
           <el-link :underline="false" type="primary" @click="openDetail(scope.row.id)">
             {{ scope.row.name }}
@@ -99,10 +111,25 @@
         </template>
       </el-table-column>
       <el-table-column :label="t('customer.mobile')" align="center" prop="mobile" min-width="120" />
-      <el-table-column :label="t('customer.telephone')" align="center" prop="telephone" min-width="130" />
+      <el-table-column
+        :label="t('customer.telephone')"
+        align="center"
+        prop="telephone"
+        min-width="130"
+      />
       <el-table-column :label="t('customer.email')" align="center" prop="email" min-width="180" />
-      <el-table-column :label="t('customer.detailAddress')" align="center" prop="detailAddress" min-width="180" />
-      <el-table-column align="center" :label="t('customer.industryId')" prop="industryId" min-width="100">
+      <el-table-column
+        :label="t('customer.detailAddress')"
+        align="center"
+        prop="detailAddress"
+        min-width="180"
+      />
+      <el-table-column
+        align="center"
+        :label="t('customer.industryId')"
+        prop="industryId"
+        min-width="100"
+      >
         <template #default="scope">
           <dict-tag :type="DICT_TYPE.CRM_CUSTOMER_INDUSTRY" :value="scope.row.industryId" />
         </template>
@@ -127,9 +154,24 @@
         :formatter="dateFormatter"
         min-width="180"
       />
-      <el-table-column align="center" :label="t('clue.contactLastContent')" prop="contactLastContent" min-width="200" />
-      <el-table-column align="center" :label="t('clue.ownerUserName')" prop="ownerUserName" min-width="100" />
-      <el-table-column align="center" :label="t('clue.ownerUserDeptName')" prop="ownerUserDeptName" min-width="100" />
+      <el-table-column
+        align="center"
+        :label="t('clue.contactLastContent')"
+        prop="contactLastContent"
+        min-width="200"
+      />
+      <el-table-column
+        align="center"
+        :label="t('clue.ownerUserName')"
+        prop="ownerUserName"
+        min-width="100"
+      />
+      <el-table-column
+        align="center"
+        :label="t('clue.ownerUserDeptName')"
+        prop="ownerUserDeptName"
+        min-width="100"
+      />
       <el-table-column
         :label="t('clue.updateTime')"
         align="center"
@@ -144,25 +186,43 @@
         :formatter="dateFormatter"
         min-width="180"
       />
-      <el-table-column align="center" :label="t('clue.creatorName')" prop="creatorName" min-width="100" />
-      <el-table-column :label="t('common.action')" align="center" min-width="150" fixed="right">
+      <el-table-column
+        align="center"
+        :label="t('clue.creatorName')"
+        prop="creatorName"
+        min-width="100"
+      />
+      <el-table-column :label="t('common.action')" align="center" fixed="right" width="140">
         <template #default="scope">
-          <el-button
-            link
-            type="primary"
-            @click="openForm('update', scope.row.id)"
-            v-hasPermi="['crm:clue:update']"
-          >
-            {{ t('common.edit') }}
-          </el-button>
-          <el-button
-            link
-            type="danger"
-            @click="handleDelete(scope.row.id)"
-            v-hasPermi="['crm:clue:delete']"
-          >
-            {{ t('common.delete') }}
-          </el-button>
+          <TableActions mode="menu">
+            <el-button
+              v-if="!scope.row.transformStatus && scope.row.poolStatus === 0"
+              link
+              type="primary"
+              @click="openForm('update', scope.row.id)"
+              v-hasPermi="['crm:clue:update']"
+            >
+              {{ t('common.edit') }}
+            </el-button>
+            <el-button
+              v-if="!scope.row.transformStatus && scope.row.poolStatus === 0"
+              link
+              type="danger"
+              @click="handleDelete(scope.row.id)"
+              v-hasPermi="['crm:clue:delete']"
+            >
+              {{ t('common.delete') }}
+            </el-button>
+            <el-button
+              v-if="!scope.row.transformStatus && scope.row.poolStatus === 0"
+              link
+              type="warning"
+              @click="handlePutPublic(scope.row)"
+              v-hasPermi="['crm:clue-public:put']"
+            >
+              {{ t('clue.putPublic') }}
+            </el-button>
+          </TableActions>
         </template>
       </el-table-column>
     </el-table>
@@ -183,6 +243,7 @@
 import { DICT_TYPE } from '@/utils/dict'
 import { dateFormatter } from '@/utils/formatTime'
 import download from '@/utils/download'
+import { resolveDialogAction } from '@/utils/dialogAction'
 import * as ClueApi from '@/api/crm/clue'
 import ClueForm from './ClueForm.vue'
 import { TabsPaneContext } from 'element-plus'
@@ -233,7 +294,7 @@ const resetQuery = () => {
 
 /** tab 切换 */
 const handleTabClick = (tab: TabsPaneContext) => {
-  queryParams.sceneType = tab.paneName
+  queryParams.sceneType = String(tab.paneName)
   handleQuery()
 }
 
@@ -251,27 +312,40 @@ const openForm = (type: string, id?: number) => {
 
 /** 删除按钮操作 */
 const handleDelete = async (id: number) => {
-  try {
-    // 删除的二次确认
-    await message.delConfirm()
-    // 发起删除
-    await ClueApi.deleteClue(id)
-    message.success(t('common.delSuccess'))
-    // 刷新列表
-    await getList()
-  } catch {}
+  if (!(await resolveDialogAction(message.delConfirm()))) return
+  await ClueApi.deleteClue(id)
+  message.success(t('common.delSuccess'))
+  await getList()
+}
+
+/** 放入公共线索池 */
+const handlePutPublic = async (row: ClueApi.ClueVO) => {
+  const result = await resolveDialogAction(
+    ElMessageBox.prompt(
+      t('clue.putPublicReasonPrompt', { name: row.name }),
+      t('clue.putPublic'),
+      {
+        inputValidator: (input: string) => {
+          if (!input?.trim()) return t('clue.putPublicReasonRequired')
+          if (input.trim().length > 500) return t('clue.putPublicReasonTooLong')
+          return true
+        }
+      }
+    )
+  )
+  if (!result) return
+  await ClueApi.putCluePublic({ clueId: row.id, reason: result.value.trim() })
+  message.success(t('clue.putPublicSuccess'))
+  await getList()
 }
 
 /** 导出按钮操作 */
 const handleExport = async () => {
+  if (!(await resolveDialogAction(message.exportConfirm()))) return
+  exportLoading.value = true
   try {
-    // 导出的二次确认
-    await message.exportConfirm()
-    // 发起导出
-    exportLoading.value = true
     const data = await ClueApi.exportClue(queryParams)
     download.excel(data, t('clue.exportFileName') + '.xls')
-  } catch {
   } finally {
     exportLoading.value = false
   }
